+++ Sources/Shared/Publication/Contributor.swift
@@ -17,6 +17,10 @@ public struct Contributor: Hashable, Sendable, JSONValueDecodable, JSONObjectEnc
     /// An unambiguous reference to this contributor.
     public var identifier: String?
 
+    /// Alternate identifiers for this contributor, other than the primary
+    /// ``identifier``.
+    public var altIdentifiers: [AltIdentifier]
+
     /// The string used to sort the name of the contributor.
     public var sortAs: String?
 
@@ -32,6 +36,7 @@ public struct Contributor: Hashable, Sendable, JSONValueDecodable, JSONObjectEnc
     public init(
         name: LocalizedStringConvertible,
         identifier: String? = nil,
+        altIdentifiers: [AltIdentifier] = [],
         sortAs: String? = nil,
         roles: [String] = [],
         role: String? = nil,
@@ -46,6 +51,7 @@ public struct Contributor: Hashable, Sendable, JSONValueDecodable, JSONObjectEnc
 
         localizedName = name.localizedString
         self.identifier = identifier
+        self.altIdentifiers = altIdentifiers
         self.sortAs = sortAs
         self.roles = roles
         self.position = position
@@ -67,6 +73,7 @@ public struct Contributor: Hashable, Sendable, JSONValueDecodable, JSONObjectEnc
             self.init(
                 name: name,
                 identifier: dict["identifier"]?.string,
+                altIdentifiers: dict["altIdentifier"]?.decode(allowingSingle: true, warnings: warnings) ?? [],
                 sortAs: dict["sortAs"]?.string,
                 roles: dict["role"]?.decode(allowingSingle: true) ?? [],
                 position: dict["position"]?.double,
@@ -82,6 +89,7 @@ public struct Contributor: Hashable, Sendable, JSONValueDecodable, JSONObjectEnc
         .init([
             "name": localizedName,
             "identifier": identifier,
+            "altIdentifier": altIdentifiers.orNullIfEmpty,
             "sortAs": sortAs,
             "role": roles.orNullIfEmpty,
             "position": position,